Only fetch the given branch.
Introduces a new function git_fetch_branch.

When fetching a branch from a remote we only want to fetch the given
branch. Before the fetch would fetch the remote but store it in
FETCH_HEAD and do nothing with this. This meant that checks like
require_branch_equal not work correctly when given the fetch flag.