Refactor some of the branchdeleted variables.
The variables localbranchdeleted and remotebranchdeleted are not initialized causing errors on finish.

Instead of using the numbers 1 and 0, lets use the shFlags $FLAGS_TRUE and $FLAGS_FALSE. The variables used the opposite numbers for true and false in bash. In bash 0 equals true and 1 equals false, unlike most other programming languages where 0 equals false and 1 equals true.