Fix variable non-zero test
Due to missing quotation marks, the -n test won't work properly, leading
to (strangely) always returning true and overriding variables with
nothing.

That causes "release finish" and "hotfix finish" to silently ignore the
flags -u, -m, and -f, and perhaps also some others which I didn't try.

With the quotation marks in place now that shouldn't be an issue
anymore.