Fix variable assignment
The current implementation of override_flag_string did not allow strings
that contain whitespaces. For example if one sets an environment
variable like so: GITFLOW_FLAG_RELEASE_FINISH_MESSAGE="Tag release",
then a release finish, which calls the override method, would produce an
error, that the command "release" is not found.

So, when the value of the assignment is quoted, there is no problem and
it behaves as expected. (Compare assignment `var=an example` with
`var="an example"`.)