Resolve #148 - Export command fails in 'gitflow_load_settings()'
See https://github.com/petervanderdoes/gitflow/issues/148 for a full description
of this issue.  In short, the export can fail (either raising an error from the
shell, or silently exporting an incorrect value) if the path to the repo
contains spaces.

It is solved by quoting the value to be exported, preventing the shell
from splitting the result into multiple values (and then trying to parse those
as multiple variable-names & values).