Fix "[: -eq: unexpected operator" message
If the first DEFINE is not DEFINE_boolean the shell prints an error message.
Using the example in the header of gitflow-shFlags:

./hello.sh: 359: [: -eq: unexpected operator
Hello, world.

This is triggered by

       if [ ${_flags_type_} -eq ${__FLAGS_TYPE_BOOLEAN} \
          -a  ${_flags_isNegate_} -eq ${FLAGS_TRUE} ]; then

where `_flags_isNegate_` is not set and hence the shell reads `-a -eq 0`

This is masked because by the first DEFINE_boolean because `_flags_define`
does not unset the internal variable `_flags_isNegate_` at the end and hence
the value spills over to the next invocation.

Currently all commands seems to define a boolean option first, so strictly
speaking gitflow is not affected.

Signed-off-by: Andreas Heiduk <asheiduk@gmail.com>