Perform git push before deleting branch
This addresses an issue when using a feature branch that is associated
with a pull request on GitHub when the develop branch is merge
protected.  In the current implementation of `git flow feature finish`,
the remote feature branch is deleted before the merged develop branch
has been pushed.  This causes the pull request to be closed before the
changes have been merged, so the approval no longer applies.  To a
certain extent this seems to be a bug with GitHub, but I don't see a
drawback to pushing before deleting the remote branch.